add: InMemoryQueueclient
