+++ src/main/kotlin/org/wfanet/measurement/common/rabbitmq/BUILD.bazel
@@ -3,11 +3,19 @@ load("@wfa_rules_kotlin_jvm//kotlin:defs.bzl", "kt_jvm_library")
 package(default_visibility = ["//visibility:public"])
 
 kt_jvm_library(
+    name = "queue_client",
+    srcs = ["QueueClient.kt"],
+    deps = [
+        "//imports/java/com/rabbitmq/client",
+        "//imports/kotlin/kotlinx/coroutines:core",
+    ],
+)
+
+kt_jvm_library(
     name = "rabbit_mq_client",
     srcs = ["RabbitMqClient.kt"],
     deps = [
         "//imports/java/com/rabbitmq/client",
-        "//imports/java/picocli",
         "//imports/kotlin/kotlinx/coroutines:core",
         "//src/main/kotlin/org/wfanet/measurement/common",
     ],
