+++ src/main/kotlin/org/wfanet/measurement/common/rabbitmq/QueueClient.kt
@@ -1,0 +1,36 @@
+// Copyright 2024 The Cross-Media Measurement Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.wfanet.measurement.common.rabbitmq
+
+import com.rabbitmq.client.Channel
+import kotlinx.coroutines.channels.ReceiveChannel
+
+interface QueueClient : AutoCloseable {
+  fun <T> subscribe(queueName: String): ReceiveChannel<QueueMessage<T>>
+
+  data class QueueMessage<T>(
+    val body: T,
+    private val deliveryTag: Long,
+    private val channel: Channel,
+  ) {
+    fun ack() {
+      channel.basicAck(deliveryTag, false)
+    }
+
+    fun nack(requeue: Boolean = true) {
+      channel.basicNack(deliveryTag, false, requeue)
+    }
+  }
+}
